# swagger.yaml
swagger: '2.0'
info:
  description: Sample Service
  version: 1.2.3
  title: Sample Service
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0'
host: www.servicesauthlb.ford.com
basePath: /api/Prospect/v3
tags:
  - name: abc-controller
    description: Sample Controller
schemes:
  - https
paths:
  '/Lead/{campaign}/{site}/{eventType6}':
    post:
      tags:
        - abc-controller
      summary: Sample Service
      operationId: getAPIUsingPOST
      consumes:
        - application/json
      produces:
        - application/json
      parameters:
        - name: campaign
          in: path
          description: campaign
          required: true
          type: string
          pattern: '^[a-zA-Z0-9]+$'
          maxLength: 10
          x-42c-sample: EP00000011
        - name: eventType
          in: path
          description: eventType
          required: true
          type: string
          pattern: A|ML1|TDB|TDR|SLN|TDT|SLR|SL1|BRN|TDN
          maxLength: 3
          x-42c-sample: A
        - in: body
          name: form
          description: form
          required: true
          schema:
            $ref: '#/definitions/Form'
        - name: locale
          in: query
          description: locale
          required: true
          type: string
          pattern: '^[a-z]{2}_[A-Z]{2}$'
          maxLength: 5
          x-42c-sample: en_GB
        - name: site
          in: path
          description: site
          required: true
          type: string
          pattern: '^[0-9]+$'
          x-42c-sample: '555'
        - name: validateOnly
          in: query
          description: validateOnly
          required: false
          type: boolean
          x-42c-sample: false
      responses:
        '200':
          description: OK
          schema:
            $ref: '#/definitions/Envelope'
      security:
        - oauth2:
            - default
      deprecated: false
securityDefinitions:
  oauth2:
    type: oauth2
    tokenUrl: 'https://corpqa.sts.ford.com/adfs/oauth2/token'
    flow: application
    scopes:
      default: provides access to all endpoints
definitions:
  Envelope:
    type: object
    properties:
      data:
        type: object
      status:
        $ref: '#/definitions/Status'
    title: Envelope
  Event:
    type: object
    properties:
      dealerID:
        type: string
        pattern: ^(.)+$
        maxLength: 50
        x-42c-sample: 44182AA
      groupCode:
        type: string
        x-42c-sample: '1'
      itemCode:
        type: string
        x-42c-sample: '000'
      sequence:
        type: integer
        x-42c-sample: '1'
    title: Event
  Form:
    type: object
    properties:
      destination:
        type: string
        pattern: XIN|CCL|SIE
        maxLength: 3
        x-42c-sample: XIN
      events:
        type: array
        items:
          $ref: '#/definitions/Event'
        minItems: 1
        maxItems: 9
      fields:
        type: array
        items:
          $ref: '#/definitions/SubmitFields'
        minItems: 4
        maxItems: 14
      pageUrl:
        type: string
        pattern: ^(.)+$
        x-42c-sample: >-
          https://www.brandeu.ford.com/content/guxeu/uk/en_gb/home/shop/research/request-a-test-drive.html
      permission:
        type: array
        items:
          $ref: '#/definitions/Permission'
        maxItems: 5
      suppression:
        type: boolean
        x-42c-sample: true
    title: Form
  Permission:
    type: object
    properties:
      name:
        type: string
        pattern: Telephone|Email|Mobile|SMS|Address
        maxLength: 9
        x-42c-sample: Email
      value:
        type: boolean
        x-42c-sample: true
    title: Permission
  Status:
    type: object
    properties:
      errorMessage:
        type: string
      incidentId:
        type: string
      status:
        type: string
        minLength: 0
        maxLength: 20
      statusCode:
        type: integer
        format: int32
        minimum: 0
        maximum: 3
    title: Status
  SubmitFields:
    type: object
    properties:
      name:
        type: string
        pattern: >-
          Title|FirstName|Surname|Email|AddressLine1|AddressLine2|AddressLine3|AddressLine4|Town|PostCode|Comments|FreeFormText|TelephoneHome|TelephoneWork|TelephoneMobile
        maxLength: 20
        x-42c-sample: FirstName
      value:
        type: string
        x-42c-sample: Pavan
    title: SubmitFields